+++ CHANGELOG.md
@@ -7,11 +7,19 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+#### MIDI events
+
+The new functions below replace the `MIDIEvent` class and `Take.midi_events`. Those are deprecated because of the instability of index-based methods. See discussion in [#81](https://github.com/RomeoDespres/reapy/pull/81#issuecomment-775458862).
+
+- [`Take.get_midi_event`] to retrieve MIDI event data by index.
+- [`Take.get_midi_events`] to retrieve all MIDI events at once.
+- [`Take.set_midi_event`] to set MIDI event data by index.
+- [`Take.get_first_selected_event_index`] to iterate through selected MIDI events.
+
+#### Other
+
 - [`reapy.map`](https://python-reapy.readthedocs.io/en/latest/reapy.core.html#reapy.core.map) for efficient mapping of `reapy` functions to large iterables of arguments.
 - [`Take.midi_grid`] to retrieve grid properties of the most recent MIDI editor for a take.
-- [`Take.get_midi_event`] to retrieve MIDI event data by index. This function intends to replace `Take.midi_events` which is deprecated because of the instability of index-based methods. See discussion in [#81](https://github.com/RomeoDespres/reapy/pull/81#issuecomment-775458862).
-- [`Take.set_midi_event`] to set MIDI event data by index. This function intends to replace `MIDIEvent.set` which is deprecated because of the instability of index-based methods. See discussion in [#81](https://github.com/RomeoDespres/reapy/pull/81#issuecomment-775458862).
-- [`Take.get_first_selected_event_index`]
 
 ### Fixed
 
@@ -519,6 +527,7 @@ Project.play_state was buggy and has been removed.
 [`Take.get_first_selected_event_index`]: https://python-reapy.readthedocs.io/en/latest/reapy.core.html#reapy.core.Take.get_first_selected_event_index
 [`Take.get_info_value`]: https://python-reapy.readthedocs.io/en/latest/reapy.core.html#reapy.core.Take.get_info_value
 [`Take.get_midi_event`]: https://python-reapy.readthedocs.io/en/latest/reapy.core.html#reapy.core.Take.get_midi_event
+[`Take.get_midi_events`]: https://python-reapy.readthedocs.io/en/latest/reapy.core.html#reapy.core.Take.get_midi_events
 [`Take.has_valid_id`]: https://python-reapy.readthedocs.io/en/latest/reapy.core.html#reapy.core.Take.has_valid_id
 [`Take.is_midi`]: https://python-reapy.readthedocs.io/en/latest/reapy.core.html#reapy.core.Take.is_midi
 [`Take.item`]: https://python-reapy.readthedocs.io/en/latest/reapy.core.html#reapy.core.Take.item